emokit - Python 3 : (Windows + Linux) Hybrid
This is not yet functional. 

Currently, jonrei's version sets the Linux device with
headset['device'] = path
and then to close it, headset['device'].close()

however, Windows hidapi, seems to be a little different, as the
device handle doesn't use paths, so I'm still trying to figure out
how to set the device handle.

Currently, it gives the error
```
  'bool' object does not support item assignment
```

Its a little closer to what it should be, but its all a bit of a mess right now.